Removed the ConstrainedBox from the HomePage and placed a container and then converting the container into SaizedBox
