+++ lib/screens/home/category_item.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:todo_list/shared/components/progress_bar.dart';
 import 'package:velocity_x/velocity_x.dart';
 import 'package:todo_list/constantes.dart';
 
@@ -11,6 +12,9 @@ class CategoryItem extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       width: 220,
+      height: 110,
+
+      ///o height nesse caso controla a altura do box
       padding: EdgeInsets.symmetric(
           horizontal: k.defaulPadding, vertical: k.defaulPadding),
       decoration: BoxDecoration(
@@ -29,51 +33,17 @@ class CategoryItem extends StatelessWidget {
         children: [
           '40 tasks'.text.color(k.defaultGray).make(),
           'Business'.text.xl2.bold.make(),
-          ProgressBar(color: Vx.purple300),
+          SizedBox(
+            height: 12,
+          ),
+
+          ///foi criada a progressBar que fica dentro do boxItem no casoCategoryItem()
+          ProgressBar(
+            color: Vx.purple300,
+            percent: 80,
+          ),
         ],
       ),
-    );
-  }
-}
-
-class ProgressBar extends StatelessWidget {
-  final Color color;
-  const ProgressBar({
-    Key? key,
-    required this.color,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        Container(
-          height: 4,
-          width: 172,
-          decoration: BoxDecoration(
-            color: Vx.blueGray200,
-            borderRadius: BorderRadius.circular(4),
-          ),
-        ),
-        AnimatedContainer(
-          duration: Duration(milliseconds: 300),
-          curve: Curves.easeInOut,
-          height: 4,
-          width: 80,
-          decoration: BoxDecoration(
-            color: color,
-            borderRadius: BorderRadius.circular(4),
-            boxShadow: [
-              BoxShadow(
-                offset: Offset(0, 2),
-                blurRadius: 3,
-                spreadRadius: -5,
-                color: Vx.blue500.withOpacity(0.25),
-              ),
-            ],
-          ),
-        ),
-      ],
     );
   }
 }
